+++ src/openai/types/beta/threads/run_status.py
@@ -1,10 +1,10 @@
 # File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
-from typing_extensions import Literal
+from typing_extensions import Literal, TypeAlias
 
 __all__ = ["RunStatus"]
 
-RunStatus = Literal[
+RunStatus: TypeAlias = Literal[
     "queued",
     "in_progress",
     "requires_action",